Add files via upload
During my time working with the CRF-C one of my projects was to create this infographic. Firstly I had to choose which online software was the most appropriate for the task I had to complete. Once this was established I created a basic infographic as a platform to expand on. It took a few weeks to get it to the final stage which you can see I have included. Many emails back and fourth tweaking the smallest of things but we got there in the end.