Add files via upload
As part of my summer internship with the CRF-C, one of my tasks was to retrieve data from Castor through R. Castor is a cloud-based clinical data management platform. The main three components to Castor included the study, reports and surveys. The facility to retrieve survey information and results through R was possible but it was extremely difficult to make sense of the information. My task was to retrieve that information and make it presentable. I hate attached my R script and a word document which includes the code used in my R script.